+++ apps/cxc/app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/auth/login/LoginForm";
+import { LoginForm } from "@/components/login/LoginForm";
 import Image from "next/image";
 
 export default function LoginPage() {
